requirements.txt
click>=8.1.3
Flask==2.3.2
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe==2.1.2
python-dotenv==0.21.1
Werkzeug>=2.3.3
zipp==3.11.0